requirements.txt
pandas>=1.5
openpyxl>=3.0
SQLAlchemy>=1.4
mysql-connector-python>=8.0
dash>=2.0
plotly>=5.0
gunicorn>=20.1
python-dotenv>=1.0
pytest>=7.0
